Auto tests: Fix some MSVC warnings about integer conversions

+++ tests/auto/corelib/tools/collections/tst_collections.cpp
@@ -2962,7 +2962,7 @@ public:
 
     inline bool operator==(const AlignedClass &other) const { return i == other.i; }
     inline bool operator<(const AlignedClass &other) const { return i < other.i; }
-    friend inline int qHash(const AlignedClass &a) { return qHash(a.i); }
+    friend inline size_t qHash(const AlignedClass &a) { return qHash(a.i); }
 };
 
 using Aligned4 = AlignedClass<4>;
